refactor: split store.rs
